requirements.txt
Jinja2 >= 2.10.1
openstacksdk >= 0.36.4
python-openstackclient >= 4.0.1